Move version parser to a separate file
