+++ internal/nginx/version_test.go
@@ -1,20 +1,22 @@
-package nginx
+package nginx_test
 
 import (
 	"testing"
+
+	"github.com/nginxinc/kubernetes-ingress/internal/nginx"
 )
 
 func TestNginxVersionParsing(t *testing.T) {
 	t.Parallel()
 	type testCase struct {
 		input    string
-		expected Version
+		expected nginx.Version
 	}
 	testCases := []testCase{
 		{
 			input: "nginx version: nginx/1.25.1 (nginx-plus-r30-p1)",
-			expected: Version{
-				raw:    "nginx version: nginx/1.25.1 (nginx-plus-r30-p1)",
+			expected: nginx.Version{
+				Raw:    "nginx version: nginx/1.25.1 (nginx-plus-r30-p1)",
 				OSS:    "1.25.1",
 				IsPlus: true,
 				Plus:   "nginx-plus-r30-p1",
@@ -22,8 +24,8 @@ func TestNginxVersionParsing(t *testing.T) {
 		},
 		{
 			input: "nginx version: nginx/1.25.3 (nginx-plus-r31)",
-			expected: Version{
-				raw:    "nginx version: nginx/1.25.3 (nginx-plus-r31)",
+			expected: nginx.Version{
+				Raw:    "nginx version: nginx/1.25.3 (nginx-plus-r31)",
 				OSS:    "1.25.3",
 				IsPlus: true,
 				Plus:   "nginx-plus-r31",
@@ -31,8 +33,8 @@ func TestNginxVersionParsing(t *testing.T) {
 		},
 		{
 			input: "nginx version: nginx/1.25.0",
-			expected: Version{
-				raw:    "nginx version: nginx/1.25.0",
+			expected: nginx.Version{
+				Raw:    "nginx version: nginx/1.25.0",
 				OSS:    "1.25.0",
 				IsPlus: false,
 				Plus:   "",
@@ -41,7 +43,7 @@ func TestNginxVersionParsing(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.input, func(t *testing.T) {
-			actual := NewVersion(tc.input)
+			actual := nginx.NewVersion(tc.input)
 			if actual != tc.expected {
 				t.Errorf("expected %v but got %v", tc.expected, actual)
 			}
@@ -71,7 +73,7 @@ func TestNginxVersionFormat(t *testing.T) {
 	}
 	for _, tc := range tt {
 		t.Run(tc.input, func(t *testing.T) {
-			v := NewVersion(tc.input)
+			v := nginx.NewVersion(tc.input)
 			got := v.Format()
 			if got != tc.want {
 				t.Errorf("want %q but got %q", tc.want, got)
@@ -83,33 +85,33 @@ func TestNginxVersionFormat(t *testing.T) {
 func TestNginxVersionPlusGreaterThanOrEqualTo(t *testing.T) {
 	t.Parallel()
 	type testCase struct {
-		version  Version
+		version  nginx.Version
 		input    string
 		expected bool
 	}
 	testCases := []testCase{
 		{
-			version:  NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30-p1)"),
+			version:  nginx.NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30-p1)"),
 			input:    "nginx-plus-r30-p1",
 			expected: true,
 		},
 		{
-			version:  NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30)"),
+			version:  nginx.NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30)"),
 			input:    "nginx-plus-r30",
 			expected: true,
 		},
 		{
-			version:  NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30-p1)"),
+			version:  nginx.NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30-p1)"),
 			input:    "nginx-plus-r30",
 			expected: true,
 		},
 		{
-			version:  NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30)"),
+			version:  nginx.NewVersion("nginx version: nginx/1.25.1 (nginx-plus-r30)"),
 			input:    "nginx-plus-r30-p1",
 			expected: false,
 		},
 		{
-			version:  NewVersion("nginx version: nginx/1.25.1"),
+			version:  nginx.NewVersion("nginx version: nginx/1.25.1"),
 			input:    "nginx-plus-r30-p1",
 			expected: false,
 		},
